Gradio_app/requirements.txt
gradio
transformers
torch
torchvision
pillow
numpy
opencv-python
huggingface_hub
facenet-pytorch
mediapipe
